requirements.txt
retry
openai
newspaper3k
